requirements.txt
asgiref==3.5.2

certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
cryptography

asyncio==3.4.3
attrs==22.2.0
autobahn
Automat==22.10.0
boto3==1.26.99
botocore==1.29.99
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
decorator==5.1.1

drf-yasg==1.20.0
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.10
django-allauth==0.50.0

django-filter==22.1
#djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
Pillow==10.0.1
pycparser==2.21
PyJWT==2.6.0


django-filter==22.1
django-resized==1.0.2
django-rest-knox==4.2.0
djangorestframework==3.12.0
dynaconf==3.1.12
gunicorn==20.1.0
h2==4.1.0
hpack==4.0.0
hvac==1.1.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==10.0.1
priority==1.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.6.0
pyOpenSSL
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.5
requests==2.31.0
requests-oauthlib==1.3.1


sqlparse==0.4.4
urllib3==1.26.18
#psycopg2==2.9.6
django-resized==1.0.2
django-rest-knox==4.2.0
django-cors-headers==3.14.0
tzdata==2023.3

s3transfer==0.6.0
service-identity==21.1.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.18
validators==0.20.0
zope.interface==5.5.2

websocket-client==1.7.0
django-environ